+++ Sources/Filters.swift
@@ -39,7 +39,7 @@ func defaultFilter(value: Any?, arguments: [Any?]) -> Any? {
 
 func joinFilter(value: Any?, arguments: [Any?]) throws -> Any? {
   guard arguments.count < 2 else {
-    throw TemplateSyntaxError("'join' filter takes a single argument")
+    throw TemplateSyntaxError("'join' filter takes at most one argument")
   }
 
   let separator = stringify(arguments.first ?? "")
@@ -55,7 +55,7 @@ func joinFilter(value: Any?, arguments: [Any?]) throws -> Any? {
 
 func splitFilter(value: Any?, arguments: [Any?]) throws -> Any? {
   guard arguments.count < 2 else {
-    throw TemplateSyntaxError("'split' filter takes a single argument")
+    throw TemplateSyntaxError("'split' filter takes at most one argument")
   }
 
   let separator = stringify(arguments.first ?? " ")
@@ -111,3 +111,67 @@ func indent(_ content: String, indentation: String, indentFirst: Bool) -> String
   return result.joined(separator: "\n")
 }
 
+func mapFilter(value: Any?, arguments: [Any?], context: Context) throws -> Any? {
+  guard arguments.count >= 1 && arguments.count <= 2 else {
+    throw TemplateSyntaxError("'map' filter takes one or two arguments")
+  }
+
+  let attribute = stringify(arguments[0])
+  let variable = Variable("map_item.\(attribute)")
+  let defaultValue = arguments.count == 2 ? arguments[1] : nil
+
+  let resolveVariable = { (item: Any) throws -> Any in
+    let result = try context.push(dictionary: ["map_item": item]) {
+      try variable.resolve(context)
+    }
+    if let result = result { return result }
+    else if let defaultValue = defaultValue { return defaultValue }
+    else { return result as Any }
+  }
+
+
+  if let array = value as? [Any] {
+    return try array.map(resolveVariable)
+  } else {
+    return try resolveVariable(value as Any)
+  }
+}
+
+func compactFilter(value: Any?, arguments: [Any?], context: Context) throws -> Any? {
+  guard arguments.count <= 1 else {
+    throw TemplateSyntaxError("'compact' filter takes at most one argument")
+  }
+
+  guard var array = value as? [Any?] else { return value }
+
+  if arguments.count == 1 {
+    guard let mapped = try mapFilter(value: array, arguments: arguments, context: context) as? [Any?] else {
+      return value
+    }
+    array = mapped
+  }
+
+  return array.flatMap({ item -> Any? in
+    if let unwrapped = item, String(describing: unwrapped) != "nil" { return unwrapped }
+    else { return nil }
+  })
+}
+
+func filterEachFilter(value: Any?, arguments: [Any?], context: Context) throws -> Any? {
+  guard arguments.count == 1 else {
+    throw TemplateSyntaxError("'filterEach' filter takes one argument")
+  }
+
+  let attribute = stringify(arguments[0])
+  let expr = try context.environment.compileExpression(components: Token.block(value: attribute).components())
+
+  if let array = value as? [Any] {
+    return try array.filter {
+      try context.push(dictionary: ["$0": $0]) {
+        try expr.evaluate(context: context)
+      }
+    }
+  }
+
+  return value
+}